resolve duplication of quality matrices backup
The quality matrix table now uses mode,node_id, and timestamp as primary
keys. Hence the database engine will make sure, that no duplicates are
allowed.

To make sure that different independent processes use the same timestamp
we use a cron schedule for the writes, and instead of a datetime.now()
the fully deterministic scheduled timestamp is used.